disaster_assistant/requirements.txt
# requirements.txt
ultralytics==8.0.178
Pillow
transformers
torch


